+++ items.c
@@ -487,7 +487,7 @@ void do_item_stats_sizes(ADD_STAT add_stats, void *c) {
                 char key[8];
                 int klen = 0;
                 klen = snprintf(key, sizeof(key), "%d", i * 32);
-                assert(klen < sizeof(key));
+                assert(klen < (int) sizeof(key));
                 APPEND_STAT(key, "%u", histogram[i]);
             }
         }
